requirements.txt
alembic==1.3.1
Click==7.0
config==0.4.2
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
migrate==0.3.8
mysql-connector==2.2.9
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
